requirements.txt
fastapi
uvicorn
python-dotenv
openai
anthropic
pydantic
httpx
pytest
pytest-asyncio
